+++ app/Interfaces/Services/IAiPassportPhotoService.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Interfaces\Services;
+
+interface IAiPassportPhotoService
+{
+    function validatePhoto($base64Photo) : string;
+}
