+++ app/Interfaces/Services/IAuthService.php
@@ -5,8 +5,6 @@ interface IAuthService
 {
     function login($uid, $password);
 
-    function authWithIdUFFS($uid, $password);
-
     function forgotPassword(string $uid): void;
 
     function resetPassword(string $uid, string $newpassword);
