+++ app/Interfaces/Services/IIdUffsService.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Interfaces\Services;
+
+interface IIdUffsService
+{
+    function authWithIdUFFS($uid, $password);
+    function isActive($enrollment_id): bool;
+}
